% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_id.R
\name{extract_id}
\alias{extract_id}
\title{Extract levels as numeric from id column of the dataset}
\usage{
extract_id(df, col_id)
}
\arguments{
\item{df}{A dataframe}

\item{col_id}{A character value corresponding to id column name}
}
\value{
dist_id The numeric levels of id column from df
}
\description{
Extract levels as numeric from id column of the dataset
}
\examples{
extract_id(adcampaign,"id")
}
