% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_target.R
\name{extract_target}
\alias{extract_target}
\title{Extract target value of GLM}
\usage{
extract_target(formula)
}
\arguments{
\item{formula}{A character formula with target variable and predictor variables}
}
\value{
y The target variable from formula in character type
}
\description{
Extract target value of GLM
}
\examples{
extract_target("ctr~timeSlot")
}
