% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variables.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variables from GLM model}
\usage{
extract_variables(formula)
}
\arguments{
\item{formula}{A character formula with target variable and predictor variables}
}
\value{
formula_var The predictor variables from formula in formula type
}
\description{
Extract variables from GLM model
}
\examples{
extract_variables("ctr~timeSlot")
}
