% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incomplete_Loglikelihood_binomiale.R
\name{Incomplete_Loglikelihood_binomiale}
\alias{Incomplete_Loglikelihood_binomiale}
\title{Calculate the incomplete loglikelihood from mixture of binomial}
\usage{
Incomplete_Loglikelihood_binomiale(df, col_id = "id", target,
  var_weights, df_id, matrix_id, lamb, b_hk, K)
}
\arguments{
\item{df}{A dataframe}

\item{col_id}{A character value corresponding to id column name}

\item{target}{A character value corresponding to the target variable}

\item{var_weights}{A character value corresponding to the weights variable}

\item{df_id}{A list of dataframe filter by id levels}

\item{matrix_id}{A list of design matrices filter by id levels}

\item{lamb}{A numeric vector of proportion into the different clusters}

\item{b_hk}{A matrix of estimated beta}

\item{K}{A numeric value of number of clusters chosen for the mixture}
}
\value{
result A numeric value of incomplete loglikelihood
}
\description{
Calculate the incomplete loglikelihood from mixture of binomial
}
