% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_design_matrices.R
\name{init_design_matrices}
\alias{init_design_matrices}
\title{Initialize design matrices from dataframe to cluster}
\usage{
init_design_matrices(formula, df, col_id = "id")
}
\arguments{
\item{formula}{A character formula with target variable and predictor variables}

\item{df}{A dataframe to cluster}

\item{col_id}{A character value corresponding to name of id column in the dataframe df}
}
\value{
result_list A list containing the df filter by id levels, the design matrices filter by id levels, the number of rows for df filter by id levels
}
\description{
Initialize design matrices from dataframe to cluster
}
\examples{
init_design_matrices("ctr~timeSlot",adcampaign,"id")
}
