% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_lambda.R
\name{init_lambda}
\alias{init_lambda}
\title{Initialize the vector lambda of mixture proportion}
\usage{
init_lambda(K)
}
\arguments{
\item{K}{A numeric value corresponding to the number of cluster}
}
\value{
result A numeric vector of length K
}
\description{
Initialize the vector lambda of mixture proportion
}
\examples{
init_lambda(K=3)
}
