% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_subset.R
\name{init_subset}
\alias{init_subset}
\title{Initialize the estimation of beta}
\usage{
init_subset(df, K, col_id = "id")
}
\arguments{
\item{df}{A dataframe}

\item{K}{The number of dataframe to obtain depending on the number of cluster chosen for the mixture}

\item{col_id}{A character value corresponding to id column name}
}
\value{
subset_df A list of K subset of dataframe
}
\description{
Initialize the estimation of beta
}
\examples{
init_subset(adcampaign,3,"id")
}
