% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_tau.R
\name{init_tau}
\alias{init_tau}
\title{Initialize the matrix probability of each levels id to be in the clusters}
\usage{
init_tau(df, K, col_id = "id")
}
\arguments{
\item{df}{A dataframe}

\item{K}{The number of dataframe to obtain depending on the number of cluster chosen for the mixture}

\item{col_id}{A character value corresponding to id column name}
}
\value{
result_matrix A matrix of dimension : rows number is the number of cluster K, columns number is the number of distinct levels from id column
}
\description{
Initialize the matrix probability of each levels id to be in the clusters
}
\examples{
init_tau(adcampaign,3,"id")
}
