% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_density_binom.R
\name{log_density_binom}
\alias{log_density_binom}
\title{Calculate de log density of a binomial}
\usage{
log_density_binom(y, matrix_id, b_hk, k, var_weights)
}
\arguments{
\item{y}{A dataframe corresponding to a specific id levels from col_id}

\item{matrix_id}{A design matrix corresponding to a specific id levels from col_id}

\item{b_hk}{A matrix of estimated beta}

\item{k}{A numeric value to select the beta from a specific cluster}

\item{var_weights}{A character value corresponding to the weights variable}
}
\value{
res A numeric value
}
\description{
Calculate de log density of a binomial
}
