% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_BIC.R
\name{my_BIC}
\alias{my_BIC}
\title{Calculate the Bayesian Information Criterion (BIC)}
\usage{
my_BIC(nb_param, logl, nb_obs)
}
\arguments{
\item{nb_param}{The number of paramaters estimated by the EM}

\item{logl}{A numeric value which is the maximum value from Incomplete Loglikelihood}

\item{nb_obs}{A numeric value corresponding to the rows number of the whole dataframe}
}
\value{
BIC The numeric value of the BIC
}
\description{
Calculate the Bayesian Information Criterion (BIC)
}
