% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{update_beta}
\alias{update_beta}
\title{M-step : update of beta parameters}
\usage{
update_beta(formula, df, k, col_id, tau, m, w_inv, z, matrix_id)
}
\arguments{
\item{formula}{A character formula with target variable and predictor variables}

\item{df}{A dataframe}

\item{k}{The numeric value of the specific cluster to be updated}

\item{col_id}{A character value corresponding to id column name}

\item{tau}{A matrix of dimension : rows number is the number of cluster K, columns number is the number of distinct levels id}

\item{m}{A numeric iterative value}

\item{w_inv}{An inverse matrix representing the W matrix in the beta equation for the M step}

\item{z}{Working data in the EM algorithm}

\item{matrix_id}{A list of design matrices filter by id levels}
}
\value{
result_beta Estimated beta for cluster k
}
\description{
M-step : update of beta parameters
}
