% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tau.R
\name{update_tau}
\alias{update_tau}
\title{E-step : update of tau}
\usage{
update_tau(df, K, col_id = "id", beta_hk, lambda, m, df_id, n_c,
  matrix_id, var_weights, target)
}
\arguments{
\item{df}{A dataframe}

\item{K}{The numeric value of the total number of clusters}

\item{col_id}{A character value corresponding to id column name}

\item{beta_hk}{A matrix of estimated beta}

\item{lambda}{A numeric vector of proportion into the different clusters}

\item{m}{A numeric iterative value}

\item{df_id}{A list of dataframe filter by id levels}

\item{n_c}{A numeric vector containing the number of rows for each distinct id levels}

\item{matrix_id}{A list of design matrices filter by id levels}

\item{var_weights}{A character value corresponding to the weights variable}

\item{target}{A character value corresponding to the target variable}
}
\value{
result_pi Estimated probabilities of tau matrix
}
\description{
E-step : update of tau
}
