% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_w}
\alias{update_w}
\title{M-step : Update the diagonal matrix W from beta iterative equation}
\usage{
update_w(df, col_id = "id", var_weights, beta_up, df_id, matrix_id)
}
\arguments{
\item{df}{A dataframe}

\item{col_id}{A character value corresponding to id column name}

\item{var_weights}{A character value corresponding to the weights variable}

\item{beta_up}{A matrix of estimated beta in a specific cluster k}

\item{df_id}{A list of dataframe filter by id levels}

\item{matrix_id}{A list of design matrices filter by id levels}
}
\value{
omega_inv An up-to-date diagonal matrix W
}
\description{
M-step : Update the diagonal matrix W from beta iterative equation
}
